refactor eth-submit cli

+++ config/config.go
@@ -27,8 +27,8 @@ type Config struct {
 		SafeAddress    string `required:"true" default:"" json:"safeAddress" form:"safeAddress" query:"safeAddress"`
 		BridgeAddress  string `required:"true" default:"" json:"bridgeAddress" form:"bridgeAddress" query:"bridgeAddress"`
 		PrivateKey     string `required:"true" default:"" json:"privateKey" form:"privateKey" query:"privateKey"`
-		MaxGasFee      int    `required:"true" default:"30" json:"maxGasFee" form:"maxGasFee" query:"maxGasFee"`
-		MaxPriorityFee int    `required:"true" default:"2" json:"maxPriorityFee" form:"maxPriorityFee" query:"maxPriorityFee"`
+		MaxGasFee      int64  `required:"true" default:"30" json:"maxGasFee" form:"maxGasFee" query:"maxGasFee"`
+		MaxPriorityFee int64  `required:"true" default:"2" json:"maxPriorityFee" form:"maxPriorityFee" query:"maxPriorityFee"`
 	}
 }
